modesetting/pageflip: make do_queue_flip_on_crtc() more generic
This function was really just a way to cast one variable and was used
in a single function, which was itself doing the flush_drm_event on
error.

With this patch, this function is now hiding the nasty loop into it,
to make it easy to re-use. We'll use it in a later patch.

Signed-off-by: Martin Peres <martin.peres@linux.intel.com>